+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Database migrations
+* Improved GitRepo & BuildLog API
+    * Pagination using `limit` & `offset` query params
+    * GitRepo: filter by repo
+    * BuildLog: filter by start & end date, repo, exit code & arch
+* CLI flags to take advantage of above API improvements
 
 ## [0.3.0-alpha.2](https://git.rustybever.be/vieter/vieter/src/tag/0.3.0-alpha.2)